Moving all functions related to af_array handling to api/c/array.cpp
